pyproject.toml
[project]
name = "llm-chatbot"
version = "0.1.0"
description = "A Haystack 2.x Ollama based chatbot for user assistance interpreting manuals"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Alex8642"}
]

# Core dependencies required for the application
dependencies = [
    # API and Server
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
    
    # Haystack and Vector Search
    "haystack-ai>=2.16.0",
    "ollama-haystack==5.1.0",
    "qdrant-haystack==9.2.0",
    "qdrant-client>=1.12.0",
    "opensearch-haystack==4.2.0",
    "opensearch-py==2.6.0",
    
    # Machine Learning and NLP
    "sentence-transformers>=3.0.0",
    "torch>=2.3.1",
    "transformers>=4.40.0",
    "scikit-learn>=1.5.1",
    "numpy>=1.26.4",
    "pandas==2.2.2",
    
    # Document Processing
    "pymupdf>=1.24.10",
    "unstructured>=0.15.9",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.30.1",
    "jupyter>=1.0.0",
    "black>=24.2.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "pytest>=8.0.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

# Removing unused dependencies:
# - elasticsearch-haystack (using opensearch instead)
# - milvus-haystack (using qdrant instead)
# - weaviate-haystack (using qdrant instead)
# - npm (frontend has its own package.json)
